refactor(arch): Rename background_processes to sources, closes #201
