+++ tests/test_fetcher.py
@@ -30,7 +30,7 @@ def test_fetch_previous_borrower_awards_empty(start_background_db, caplog):  # n
         with common_test_functions.mock_response(
             200,
             [],
-            "app.background_processes.colombia_data_access.get_previous_contracts",
+            "app.sources.colombia.get_previous_contracts",
         ):
             background.fetch_previous_awards(models.Borrower(**borrower_result), common_test_client.get_test_db)
 
@@ -42,7 +42,7 @@ def test_fetch_previous_borrower_awards(start_background_db, caplog):  # noqa
         with common_test_functions.mock_response(
             200,
             contracts,
-            "app.background_processes.colombia_data_access.get_previous_contracts",
+            "app.sources.colombia.get_previous_contracts",
         ):
             background.fetch_previous_awards(models.Borrower(**borrower_result), common_test_client.get_test_db)
 
@@ -51,9 +51,7 @@ def test_fetch_previous_borrower_awards(start_background_db, caplog):  # noqa
 
 def test_fetch_empty_contracts(start_background_db, caplog):  # noqa
     with caplog.at_level("INFO"):
-        with common_test_functions.mock_response(
-            200, [], "app.background_processes.colombia_data_access.get_new_contracts"
-        ):
+        with common_test_functions.mock_response(200, [], "app.sources.colombia.get_new_contracts"):
             fetch_awards()
 
     assert "No new contracts" in caplog.text
@@ -65,13 +63,13 @@ def test_fetch_new_awards_borrower_declined(client, mock_templated_email):  # no
     with common_test_functions.mock_response_second_empty(
         200,
         contract,
-        "app.background_processes.colombia_data_access.get_new_contracts",
+        "app.sources.colombia.get_new_contracts",
     ), common_test_functions.mock_whole_process_once(
         200,
         award,
         borrower,
         email,
-        "app.background_processes.background_utils.make_request_with_retry",
+        "app.sources.make_request_with_retry",
     ):
         fetch_awards()
 
@@ -80,7 +78,7 @@ def test_fetch_new_awards_from_date(start_background_db, mock_templated_email): 
     with common_test_functions.mock_response_second_empty(
         200,
         contracts,
-        "app.background_processes.colombia_data_access.get_new_contracts",
+        "app.sources.colombia.get_new_contracts",
     ), common_test_functions.mock_function_response(
         "test_hash_12345678",
         "app.util.get_secret_hash",
@@ -89,7 +87,7 @@ def test_fetch_new_awards_from_date(start_background_db, mock_templated_email): 
         award,
         borrower,
         email,
-        "app.background_processes.background_utils.make_request_with_retry",
+        "app.sources.make_request_with_retry",
     ):
         fetch_awards()
